refactor(Friends): Move format timestamp timeago to utils, add translate on requested and sent friend rquests
