+++ ui/locales/en-US/main.ftl
@@ -39,6 +39,8 @@ friends = Friends
     .incoming_requests = Incoming Requests
     .outgoing_requests = Outgoing Requests
     .cancel = Cancel Request
+    .requested = Requested 
+    .sent = Sent
     .requests = Requests
     .pending = Pending
     .blocked = Blocked
