+++ ui/locales/pt-BR/main.ftl
@@ -40,6 +40,8 @@ friends = Friends
     .incoming_requests = Solicitações recebidas
     .outgoing_requests = Solicitações enviadas
     .cancel = Cancelar solicitação
+    .requested = Recebida 
+    .sent = Enviada
     .requests = Solicitações
     .pending = Pendente
     .blocked = Bloqueado
