+++ ui/src/components/friends/incoming_requests/mod.rs
@@ -1,5 +1,7 @@
+use chrono::{Utc, Duration};
 use dioxus::prelude::*;
 use kit::{elements::label::Label, components::{context_menu::{ContextMenu, ContextItem}, user_image::UserImage, indicator::{Platform, Status}}, icons::Icon};
+use rand::Rng;
 
 use crate::{state::{State, Action}, utils::language::get_local_text, components::friends::friend::{Friend, Relationship}};
 
@@ -15,12 +17,13 @@ pub fn PendingFriends(cx: Scope) -> Element {
                 text: get_local_text("friends.incoming_requests"),
             },
             friends_list.into_iter().map(|friend| {
+                let mut rng = rand::thread_rng();
                 let did = friend.did_key().clone();
                 let did_suffix: String = did.to_string().chars().rev().take(6).collect();
                 let friend_clone = friend.clone();
                 let friend_clone_clone = friend.clone();
                 let friend_clone_clone_clone = friend.clone();
-
+            
                 rsx!(
                     ContextMenu {
                         id: format!("{}-friend-listing", did),
@@ -45,6 +48,7 @@ pub fn PendingFriends(cx: Scope) -> Element {
                                 sent_friend_request: false,
                                 blocked: false,
                             },
+                            request_datetime: Utc::now() - Duration::days(rng.gen_range(0..30)),
                             user_image: cx.render(rsx! (
                                 UserImage {
                                     platform: Platform::Desktop,
